requirements.txt
appdirs==1.4.4
distlib==0.3.1
filelock==3.0.12
keyboard==0.13.5
mysql-connector-python==8.0.23
mysqlclient==2.0.3
pbr==5.5.1
protobuf==3.14.0
six==1.15.0
stevedore==3.3.0
style==1.1.0
update==0.0.1
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
